--- src/screens/Instructor/CourseSettings/Components/UploadButton/index.scss
@@ -1,15 +1,0 @@
-.email-list-uploadbtw-container {
-    .email-list-uploadbtw-example {
-        margin-top: 1em;
-        color: rgb(64, 126, 126);
-        background: rgb(245, 245, 245);
-        border-radius: 10px;
-        padding: 15px 10px 10px 10px;
-        display: flex;
-        flex-direction: column;
-    }
-    .email-list-uploadbtw {
-        width: 100%;
-        background: rgb(224, 225, 226);
-    }
-}
